// src/components/card/enum.ts
export enum ActionState {
	PRIMARY = "bg-primary-400 rounded flex justify-center items-center text-white font-normal text-labels my-1",
	SECONDARY = "bg-white rounded border-primary-400 border flex justify-center items-center text-primary-400 font-normal text-labels my-1",
	TERTIARY = "bg-neutral-100 rounded flex justify-center items-center text-neutral-300 font-normal text-labels my-1",
}

export enum TagState {
	PRIMARY = "bg-primary-100 border-primary-400 border rounded flex justify-center items-center text-primary-400 font-medium text-labels my-1",
	SECONDARY = "bg-white rounded border-neutral-100 border flex justify-center items-center text-neutral-400 font-medium text-labels my-1",
	TERTIARY = "bg-neutral-100 rounded flex justify-center items-center text-neutral-300 font-medium text-labels my-1",
}